Added new input args reaction_solvent and explicit_h_solvent
New attributes `reaction_solvent` and `explicit_h_solvent` are
added to the class `TrainArgs` for more input argument options.
`reaction_solvent` is used to specify whether the model is for
a reaction in a solvent. `explicit_h_solvent` is used to specify
whether explicit H is used for solvent molecules.

If `reaction_solvent` is True, MPNN layer will be adjusted so that one
MPNN takes reactions as input instead of molecules and another MPNN
takes solvent molecules as input. And `number_of_molecules` will be
automatically set to 2 if `reaction_solvent` is True.

`explicit_h_solvent` option allows the user to use different
explicit H options for reactions and solvents when `reaction_solvent`
is used. If `explicit_h_solvent` is True while `reaction_solvent` is
False, it will raise ValueError.