Add `save_validation_scores` and `save_train_scores` args options
Two new input arguments are added: `save_validation_scores` and
`save_train_scores`. These are both Boolean, and if they are used, the
final validation scores and final training scores are displayed in the
log files and saved as "val_scores.csv" and "train_scores.csv",
respectively.

These options can be only used for `run_training` and cannot be used for
`run_sklearn`.