Optional atom features are added
Six optional atom features are added. If each atom feature is set to
True, it will be used as an additional atom feature. The six optional
atom features are as follow:

1. atom_feature_radical_elec: the number of radical electron. One-hot
encoding [0, 1, 2, 3].

2. atom_feature_ring_size: the size of the smallest ring that the atom
is in. One-hot encoding [0, 3, 4, 5, 6, 7, 8].

3. atom_feature_lone_pair: the number of lone electron pairs. One-hot
encoding [0, 1, 2, 3, 4]

4. atom_feature_H_bond_donor: type of hydrogen bond donor. One-hot
encoding [0, 1, 2, 3], where 1-3 for hydrogen bonded to N, O and F,
respectively, and 0 if else.

5. atom_feature_H_bond_acceptor: type of hydrogen bond acceptor.
One-hot encoding [0, 1, 2, 3], where 1-3 for N, O and F with lone pairs,
respectively, 0 if else.

6. atom_feature_electronegativity: electronegativity of an atom. Float.